fix: 고객센터 게시글 OPEN 상태가 아닐 때 삭제 가능한 버그 수정

+++ src/main/java/com/jay/auth/service/SupportPostService.java
@@ -111,6 +111,10 @@ public class SupportPostService {
             throw new SupportPostAccessDeniedException();
         }
 
+        if (post.getStatus() != PostStatus.OPEN) {
+            throw new SupportPostNotModifiableException();
+        }
+
         supportCommentRepository.deleteByPostId(postId);
         supportPostRepository.delete(post);
